+++ payments-core/src/main/java/com/stripe/android/model/RadarSessionWithHCaptcha.kt
@@ -1,0 +1,13 @@
+package com.stripe.android.model
+
+import androidx.annotation.RestrictTo
+import com.stripe.android.core.model.StripeModel
+import kotlinx.parcelize.Parcelize
+
+@Parcelize
+@RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
+data class RadarSessionWithHCaptcha(
+    val id: String,
+    val passiveCaptchaSiteKey: String?,
+    val passiveCaptchaRqdata: String?
+) : StripeModel