+++ paymentsheet/src/main/java/com/stripe/android/common/ui/PrimaryButton.kt
@@ -3,8 +3,8 @@ package com.stripe.android.common.ui
 import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.BoxWithConstraints
+import androidx.compose.foundation.layout.defaultMinSize
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.shape.RoundedCornerShape
@@ -21,9 +21,11 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.res.dimensionResource
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.unit.dp
+import com.stripe.android.paymentsheet.R
 import com.stripe.android.uicore.StripeTheme
 import com.stripe.android.uicore.getBackgroundColor
 import com.stripe.android.uicore.getBorderStrokeColor
@@ -65,7 +67,9 @@ internal fun PrimaryButton(
                 onClick = onButtonClick,
                 modifier = Modifier
                     .fillMaxWidth()
-                    .height(44.dp),
+                    .defaultMinSize(
+                        minHeight = dimensionResource(id = R.dimen.stripe_paymentsheet_primary_button_height)
+                    ),
                 enabled = isEnabled,
                 shape = shape,
                 border = borderStroke,