+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # CHANGELOG
 
 ## XX.XX.XX - 2023-XX-XX
+* [CHANGED][7722](https://github.com/stripe/stripe-android/pull/7722) Added passive hcaptcha on radar session creation endpoint. In order to enable hCaptcha on this endpoint, you must provide an activity to this method call (see the new optional activity parameter).
+
+Dependencies updated:
+* [7603](https://github.com/stripe/stripe-android/pull/7603) Bumped compile SDK from 33 to 34.
+
+## 20.36.1 - 2024-01-08
+
+### Identity
+* [FIXED][7757](https://github.com/stripe/stripe-android/pull/7757) Fixed a crash when retrying the selfie scan.
 
 ## 20.36.0 - 2023-12-18